/usr/bin/bash to /bin/bash if [ -f "$script_name" ]; then check

+++ update_pass.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/bash
+#!/bin/bash
 now=`date +%Y%M%d%H%m%S`
 #old_pwd - old password, it could be provided
 # as the first incoming parameter to the script
@@ -7,7 +7,7 @@ old_pwd=$1
 # as the second incoming parameter to the script
 new_pwd=$2
 #file_list - the list of files in /u/debit/lib/ to be updated
-file_list=".dbpasswd_credrules .dbpasswd .credDbPasswd .dbpasswd_debit1"
+file_list=".dbpasswd_pair2 .dbpasswd_fincredrules .newdbpasswd_credrules .dbpasswd_credrules_debitapp03 .dbpasswd_credrules_aws_prod .dbpasswd_aws_prod .credDbPasswd .dbpasswd_credrules_aws11 .dbpasswd_aws11 .credDbPasswd_aws11 .dbpasswd_credrules .dbpasswd"
 #work_folder - the path to the folder for backuping initial scripts
 work_folder=/u/debit/lib
 #encrypt_key - ENCRYPTION_KEY
@@ -15,15 +15,17 @@ encrypt_key=/u/debit/lib
 
 for script_name in $work_folder/${file_list}
 do
-  echo "Start with File:       ${script_name}"
-  echo "Copying script to backup file ..."
-  cp ${script_name} ${script_name}_backup_${now}
-  echo "Changing password in script"
+  if [ -f "$script_name" ]; then
+    echo "Start with File:       ${script_name}"
+    echo "Copying script to backup file ..."
+    cp ${script_name} ${script_name}_backup_${now}
+    echo "Changing password in script"
 #decrypt old file and put its content to temporary file
-  crypt $encrypt_key < ${script_name}_backup_${now} > ./temporary_file
+    crypt $encrypt_key < ${script_name}_backup_${now} > ./temporary_file
 #old_pwd - is an old password, change it to the new one
-  sed -i "s/${old_pwd}/${new_pwd}/g" ./temporary_file
+    sed -i "s/${old_pwd}/${new_pwd}/g" ./temporary_file
 #encrypt new password
-  cat ./temporary_file | crypt --encrypt $encrypt_key > ${script_name}
-  rm ./temporary_file
+    cat ./temporary_file | crypt --encrypt $encrypt_key > ${script_name}
+    rm ./temporary_file
+  fi
 done
